requirements.txt
compel==2.0.3
dearpygui==2.0.0
diffusers==0.19.3
huggingface_hub==0.14.1
imagesize==1.4.1
numpy==1.23.5
omegaconf==2.2.3
Pillow==11.0.0
pyperclip==1.9.0
requests==2.31.0
torch==2.0.1
accelerate==0.19.0
safetensors==0.3.1
realesrgan==0.3.0
transformers==4.29.2